Move color snap logic from Note to DrawableNote
